Created modals for sign up and log in
I have created the sign up and log in modals. When you press the buttons, sign up and log in their modals will appear. the splash page is has most of its components, there's styling that needs to be fixed as i need to change the color of the buttons to the color palette